BREAKING: Fire onDragEnter, onDragLeave only once
Remove the janky timeout-based approach for detecting when drag starts and ends.

You can now rely on the dragEnter and dragLeave events to fire only once when drag starts, and once when it stops. Events originating from child nodes will be ignored correctly.

Another improvement - The dragEnter and dragLeave events will not fire if the event is not handleable by the registered onDrop or onDropText listeners, i.e. it's the wrong kind of drop.